# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyChatbot World API
  version: 1.0.0
  description: |
    MyChatbot World 플랫폼의 전체 REST API 명세.

    ## 인증
    대부분의 엔드포인트는 Supabase 발급 JWT Bearer Token 인증이 필요합니다.
    로그인 후 수신한 `access_token`을 `Authorization: Bearer {token}` 헤더에 포함하세요.

    ## 공통 에러 형식
    ```json
    { "error": { "code": "ERROR_CODE", "message": "에러 설명", "details": {} } }
    ```
  contact:
    name: MyChatbot World Support
    url: https://mychatbot-world.vercel.app
servers:
  - url: https://mychatbot-world.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth에서 발급된 JWT 액세스 토큰
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: AUTH_001
            message:
              type: string
              example: 인증이 필요합니다.
            details:
              type: object
          required:
            - code
            - message
...
paths:
...
  /api/kb/embed:
    post:
      summary: KB 임베딩 생성 (POST)
      description: KB 항목을 벡터로 변환하여 RAG 검색에 사용할 수 있도록 합니다.
      operationId: embedKb
      tags:
        - Knowledge Base
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kb_id
              properties:
                kb_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: 임베딩 생성 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  chunk_count:
                    type: integer
                    example: 12
        '500':
          description: 임베딩 생성 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Credits / Payments ────────────────────────
...
  /api/settings:
    get:
      summary: 봇 설정 조회 (GET)
      operationId: getBotSettings
      tags:
        - Settings
      security:
        - BearerAuth: []
      parameters:
        - name: chatbot_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 봇 설정 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  chatbot_id:
                    type: string
                    format: uuid
                  persona:
                    type: string
                  greeting:
                    type: string
                  model:
                    type: string
                    example: gpt-4o-mini
                  temperature:
                    type: number
                    minimum: 0.0
                    maximum: 2.0
                    example: 0.7
                  max_tokens:
                    type: integer
                    example: 1000
                  language:
                    type: string
                    example: ko
                  fallback_message:
                    type: string
                  use_kb:
                    type: boolean
                  kb_top_k:
                    type: integer
                    example: 3

    patch:
      summary: 봇 설정 업데이트 (PATCH)
      description: 봇 설정을 부분적으로 업데이트합니다 (upsert).
      operationId: updateBotSettings
      tags:
        - Settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chatbot_id
              properties:
                chatbot_id:
                  type: string
                  format: uuid
                persona:
                  type: string
                greeting:
                  type: string
                model:
                  type: string
                temperature:
                  type: number
                max_tokens:
                  type: integer
                language:
                  type: string
                fallback_message:
                  type: string
                use_kb:
                  type: boolean
                kb_top_k:
                  type: integer
      responses:
        '200':
          description: 설정 업데이트 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...